Migrate Slack access plugin tests
